디렉토리 구조 및 dummy 값 정리
